Restore behavior where hosted mode linking will not overwrite a newer compiled selection script.
And generally, older resources won't overwrite newer ones.

Review by: spoon (desk)

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@6912 8db76d5a-ed1c-0410-87a9-c151d255dfc7

+++ dev/core/src/com/google/gwt/dev/util/OutputFileSetOnDirectory.java
@@ -15,6 +15,7 @@
  */
 package com.google.gwt.dev.util;
 
+import com.google.gwt.dev.util.NullOutputFileSet.NullOutputStream;
 import com.google.gwt.dev.util.collect.HashSet;
 
 import java.io.File;
@@ -42,14 +43,28 @@ public class OutputFileSetOnDirectory extends OutputFileSet {
   }
 
   @Override
-  public OutputStream openForWrite(String path, long lastModifiedTime)
+  public OutputStream openForWrite(String path, final long lastModifiedTime)
       throws IOException {
+    final File file = makeFileForPath(path);
+    if (file.exists() && file.lastModified() >= lastModifiedTime) {
+      return new NullOutputStream();
+    }
+    mkdirs(file.getParentFile());
+    return new FileOutputStream(file) {
+      @Override
+      public void close() throws IOException {
+        super.close();
+        file.setLastModified(lastModifiedTime);
+      }
+    };
+  }
+
+  private File makeFileForPath(String path) {
     File file = dir;
     for (String part : (prefix + path).split("/")) {
       file = new File(file, part);
     }
-    mkdirs(file.getParentFile());
-    return new FileOutputStream(file);
+    return file;
   }
 
   /**
